gpu_requirements.txt
-f https://storage.googleapis.com/jax-releases/jax_releases.html
jax[cuda]==0.3.4
flax==0.4.0
optax==0.1.1
distrax==0.1.1
brax==0.0.12
tqdm
cloudpickle==2.0.0
dill
ml_collections
wandb==0.12.7
scikit-image==0.19.2
gcsfs==2022.02.0
-f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
torch==1.8.2+cpu
torchvision==0.9.2+cpu
google-auth-oauthlib==0.5
tensorflow-cpu==2.8.0
